% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_funcs.R
\name{posterior_stat_points_dgp}
\alias{posterior_stat_points_dgp}
\title{Sample from Stationary Point Univariate Posterior for DGP Model}
\usage{
posterior_stat_points_dgp(
  X,
  Y,
  num_stationary,
  total_samples,
  beta_shape1 = 1,
  beta_shape2 = 1
)
}
\arguments{
\item{X}{length n vector, x points (input data)}

\item{Y}{length n vector, y points (output data)}

\item{num_stationary}{positive integer, number of stationary points}

\item{total_samples}{positive integer, number of posterior samples to obtain}

\item{beta_shape1}{positive real number, first shape parameter for beta prior}

\item{beta_shape2}{positive real number, second shape parameter for beta prior}
}
\value{
A list with components
\describe{
  \item{dgp_stat_points}{List of length \code{num_stationary}, posterior samples for
  each stationary point}
  \item{dgp_post_var}{Vector of length \code{num_stationary}, posterior variance for
  each stationary point}
  \item{dgp_map_stat_points}{Vector of length \code{num_stationary}, MAP of stationary
  points}
  \item{intervals_list_dgp}{List of matrices, 90, 95, 99 and Bonferonni Corrected Intervals for
  each stationary point}
  \item{post_prob_gp}{Length 5000 vector from of univariate posterior applied to
   min(X) to max(X)}
}
}
\description{
Sample from the stationary point's posterior for the derivative constrained
Gaussian process method
}
