% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{posterior_stat_points_diff_mcmc}
\alias{posterior_stat_points_diff_mcmc}
\title{Sample from Stationary Point Joint Posterior for Diffeomorphism Model}
\usage{
posterior_stat_points_diff_mcmc(
  X,
  Y,
  num_betas,
  num_stationary,
  first_direction,
  zero_is_zero,
  interpolation,
  b_vec,
  prior_mean,
  prior_sd,
  n_chains,
  n_per_chain,
  cov_scale,
  apply_sm_correction = TRUE
)
}
\arguments{
\item{X}{length n vector, x points (input data)}

\item{Y}{length n vector, y points (output data)}

\item{num_betas}{positive integer > 1, number of weight parameters for the diffeomorphism}

\item{num_stationary}{positive integer, number of stationary points}

\item{first_direction}{1 or -1, 1 if the function is increasing between min(X) and the first
stationary point, -1 if the function is decreasing}

\item{zero_is_zero}{boolean, TRUE if f(min(X)) = 0, FALSE otherwise and f(min(X)) needs to be
estimated}

\item{interpolation}{'cubic' or 'linear', referring to interpolation type of lambda height vector}

\item{b_vec}{NULL or vector of length num_stationary + 2, nodes of template function. If NULL, nodes
will be equally spaced from each other}

\item{prior_mean}{(num_betas + num_stationary + 3 (2 height vectors at boundaries and for sigma)) dimensional vector of prior mean (Joint Independent Normal)}

\item{prior_sd}{(num_betas + num_stationary + 3 (2 height vectors at boundaries and for sigma)) dimension vector of prior standard deviations (Joint Independent Normal)}

\item{n_chains}{positive integer, number of MCMC chains}

\item{n_per_chain}{positive integer, number of samples per MCMC chain}

\item{cov_scale}{positive real number, scaling factor for proposal covariance matrices of MCMC chains (should be in between 1/6 and 1, higher dimensions should have a lower cov_scale)}

\item{apply_sm_correction}{boolean, whether to filter out low probability posterior modes via softmax correction}
}
\value{
A list with the components
\describe{
  \item{diff_stat_points}{List of length \code{num_stationary}, posterior samples for each stationary point}
  \item{intervals_list_diff}{List of matrices, 90, 95, 99 and Bonferroni Corrected Intervals for each stationary point}
  \item{diff_post_cov}{Posterior covariance of stationary points}
  \item{diff_map_par}{MAP of parameters}
  \item{diff_map_stat_points}{MAP of the stationary points}
  \item{map_predictions}{Matrix with X grid and MAP estimate of f(X)}
  \item{posterior_modes}{List of posterior modes, likelihoods, inverse Hessians}
  \item{acceptance_rates}{Vector of acceptance rates for MCMC chains}
}
}
\description{
Sample from posterior of Diffeomorphism Model using multiple chains of MCMC Model, with
Laplace approximations at the posterior modes as the covariance matrices. Low probability
modes are filtered out via a softmax correction of the posterior likelihood. The posterior
stationary points are outputted
}
